backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
google-generativeai==0.8.1
supabase==2.7.4
python-jose[cryptography]==3.3.0
Pillow==10.4.0
httpx==0.27.2
aiofiles==24.1.0
